requirements.txt
fastapi
uvicorn
jinja2
requests
pillow
openpyxl
